Refactor PDF service to client-side, update imports and tests
